+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './CustomButtonsRow';
 export * from './CustomCodeEditor';
 export * from './DatasourceEditor';
 export * from './DatasourcePayloadEditor';
@@ -7,5 +8,6 @@ export * from './FormElements';
 export * from './FormElementsEditor';
 export * from './FormPanel';
 export * from './HeaderParametersEditor';
+export * from './InitialFieldsEditor';
 export * from './LayoutSectionsEditor';
 export * from './NewElement';